Stabilize platform coverage tests
