+++ src/tests/files-internals-coverage.test.ts
@@ -1,11 +1,9 @@
 import assert from "node:assert/strict";
 import { mkdtemp, rm, writeFile } from "node:fs/promises";
-import { homedir, tmpdir } from "node:os";
+import { tmpdir } from "node:os";
 import { join } from "node:path";
 import test from "node:test";
 
-import { envFileInternals } from "../config/env-file.js";
-import { loadRuntimeConfig } from "../config/runtime.js";
 import {
   createDirectoryLink,
   ensureDirectory,
@@ -13,137 +11,9 @@ import {
   readTextFileOrNull,
   writeTextFile,
 } from "../files.js";
-import { officialIndexInternals } from "../official-index.js";
-import {
-  extractRepositoryUrlFromNpmMetadata,
-  fetchNpmPackageMetadata,
-  fetchNpmPackageSearch,
-  fetchPypiPackageMetadata,
-  normalizeNpmPackageMetadata,
-  normalizePypiPackageMetadata,
-  packageRegistryInternals,
-} from "../package-registries.js";
-
-void test("core root helpers cover fallback and parser edge branches", () => {
-  const config = loadRuntimeConfig({
-    AGENT_HARNESS_HOME: "   ",
-    HOME: "   ",
-    USERPROFILE: "   ",
-  });
-  assert.equal(config.paths.homeDirectory, homedir());
-
-  assert.deepEqual(
-    envFileInternals.collectDotEnvLogicalLines('NOEQUALS"quoted"\nKEY=value'),
-    ['NOEQUALS"quoted"', "KEY=value"],
-  );
-  assert.equal(envFileInternals.parseDotEnvLine("NO_SEPARATOR"), null);
-  assert.equal(envFileInternals.parseDotEnvLine("=missing-key"), null);
-  assert.equal(
-    envFileInternals.parseQuotedDotEnvValue('"unterminated', '"'),
-    "unterminated",
-  );
-  assert.equal(envFileInternals.decodeDoubleQuotedEscape("r"), "\r");
-  assert.equal(envFileInternals.decodeDoubleQuotedEscape("\\"), "\\");
-  assert.equal(envFileInternals.decodeDoubleQuotedEscape(undefined), "");
-});
-
-void test("package and official-index helpers cover remaining normalization edges", async () => {
-  const originalUrl = globalThis.URL;
-  const badName = {
-    toString(): string {
-      throw new Error("boom");
-    },
-  } as unknown as string;
-
-  globalThis.URL = class BrokenUrl {
-    constructor() {
-      throw new Error("broken");
-    }
-  } as unknown as typeof URL;
-
-  try {
-    assert.deepEqual(await fetchNpmPackageSearch("agent"), []);
-  } finally {
-    globalThis.URL = originalUrl;
-  }
-
-  assert.equal(await fetchNpmPackageMetadata(badName), null);
-  assert.equal(await fetchPypiPackageMetadata(badName), null);
-
-  assert.equal(
-    officialIndexInternals.normalizeGitHubRepositoryUrl(
-      "https://gitlab.com/example/project",
-    ),
-    null,
-  );
-  assert.equal(
-    officialIndexInternals.normalizeGitHubRepositoryUrl("not-a-url"),
-    null,
-  );
-
-  const objectRepository = normalizeNpmPackageMetadata(
-    {
-      repository: { type: 42, url: "github:example/object-repo" },
-      time: { modified: 5 },
-    },
-    "fallback-name",
-  );
-  assert.deepEqual(objectRepository.repository, {
-    type: undefined,
-    url: "https://github.com/example/object-repo",
-  });
-  assert.equal(objectRepository.lastUpdated, undefined);
-
-  const invalidRepository = normalizeNpmPackageMetadata(
-    {
-      repository: { url: 42 },
-    },
-    "fallback-name",
-  );
-  assert.equal(invalidRepository.repository, undefined);
-
-  const missingRepositoryUrl = extractRepositoryUrlFromNpmMetadata({
-    name: "fixture",
-    repository: { type: "git" },
-  });
-  assert.equal(missingRepositoryUrl, undefined);
-
-  const pypiMetadata = normalizePypiPackageMetadata(
-    {
-      info: {
-        name: "fixture",
-        project_urls: { Docs: 42 },
-      },
-      releases: { "1.0.0": [{}] },
-    },
-    "fallback-pypi",
-  );
-  assert.equal(pypiMetadata?.info.project_urls, undefined);
-  assert.equal(pypiMetadata?.lastUpdated, undefined);
-
-  assert.deepEqual(packageRegistryInternals.normalizeStringArray("nope"), []);
-  assert.equal(
-    packageRegistryInternals.normalizeStringRecord({ count: 1 }),
-    undefined,
-  );
-  assert.equal(
-    packageRegistryInternals.isGitHubRepositoryUrl("not-a-url"),
-    false,
-  );
-  assert.equal(
-    packageRegistryInternals.buildGitHubRepositoryUrl("", "repo.git"),
-    "https://github.com//repo",
-  );
-  assert.equal(
-    packageRegistryInternals.stripUrlSuffix(
-      "https://example.com/repo.git#readme",
-    ),
-    "https://example.com/repo.git",
-  );
-});
 
 void test("file internals cover usable path, ignore handling, and scan edge branches", async () => {
-  const root = await mkdtemp(join(tmpdir(), "agent-harness-files-final4-"));
+  const root = await mkdtemp(join(tmpdir(), "agent-harness-files-internals-"));
 
   try {
     const targetDirectory = join(root, "target");
